requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
psycopg2-binary==2.9.6
numpy==1.24.3
transformers==4.45.2
pandas==2.1.1
sentence-transformers==3.2.1